requirement.txt
beautifulsoup4==4.6.0
bs4==0.0.1
certifi==2017.11.5
chardet==3.0.4
Django==1.11.7
django-common-helpers==0.9.1
django-cron==0.5.0
django-widget-tweaks==1.4.1
future==0.16.0
idna==2.6
oauthlib==2.0.6
PyJWT==1.5.3
python-openid==2.2.5
python-twitter==3.3
pytz==2017.3
requests==2.18.4
requests-oauthlib==0.8.0
six==1.11.0
social-auth-app-django==2.0.0
social-auth-core==1.2.0
urllib3==1.22
twilio==4.4.0
